forfaisal-main/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2
numpy==1.24.3
google-generativeai
anthropic